Add RouterService
