+++ .gitignore
@@ -1,1 +1,5 @@
-target
+.idea
+*.iml
+
+target/
+**/*.rs.bk
